webscraping_project/requirements.txt
Scrapy==2.11.2
pandas==2.2.2
plotly==6.0.0
openpyxl==3.1.5
mysql-connector-python==9.2.0
six==1.17.0